implementation of the simulated engine. by advancing time and delivering broadcasted messages

+++ main.go
@@ -7,7 +7,7 @@ import (
 )
 
 var matrixMode bool = true
-var NodesNumber int = 5
+var NodesNumber uint8 = 5
 
 func main() {
 	var runner coreraft.Runner
